Extract Cursor SQLite IO helpers
Move retryable initialization and Cursor store path helpers into a focused IO module while keeping sqlite-reader as the stable public facade.
